removing errors and refactoring routes
