+++ lib/main.dart
@@ -2,11 +2,10 @@ import 'package:flame/game.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 
-import 'src/OneSecondGame.dart';
+import 'src/routes/OneSecondGame.dart';
 import 'package:flame/flame.dart';
 
 //for testing
-import 'src/components/sheet_service.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 
 void main() async {
