+++ lib/src/components.dart
@@ -1,1 +1,7 @@
-export 'playArea.dart';
+
+export 'package:figureout/src/components/CircleShape.dart';
+
+
+export 'package:figureout/src/routes/AftermathScreen.dart';
+
+export 'package:figureout/src/functions/UserRemovable.dart';
